+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+## TBD
+
+### Bug fixes
+
+* Fix `NSNull` handling in `+[BugsnagError errorFromJson:]` and `+[BugsnagStackframe frameFromJson:]`.
+  [#1143](https://github.com/bugsnag/bugsnag-cocoa/pull/1143)
+  [#1138](https://github.com/bugsnag/bugsnag-cocoa/issues/1138)
+
 ## 6.10.0 (2021-06-30)
 
 ### Bug fixes